_archived/google-tasks-deprecation-2025-12-16/infrastructure/google-tasks-mcp-server/requirements.txt
fastmcp>=0.3.0
google-api-python-client>=2.108.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
